backend/services/sync/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
websockets>=12.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
